Wrapper to find Signature Exposures for a set of samples

CallSignatures() calculates copy-number signature exposures for a list of samples given their segmented copy-number profiles. This function makes use of code found in the CN-Signatures bitbucket repo. and described more robustly here https://www.nature.com/articles/s41588-018-0179-8.

Usage

CallSignatureExposures(
  copy_numbers_input,
  component_models = NULL,
  signatures = NULL,
  cores = 1,
  log_features = FALSE,
  extra_features = FALSE,
  plot_savepath = NULL,
  sigs_savepath = NULL,
  relative = FALSE,
  refgenome = "hg19"
)

Arguments

copy_numbers_input

A list of dataframes. Each dataframe should be the segmented copy-numbers of a sample.

component_models

String. The mixture models to use in determining the components present in this dataset.
Options:
'component_models/30kb_ovarian/component_models_britroc_aCNs.rds' - Components determined by modelling the Britroc Ovarian sWGS samples. (Macintyre 2018)
'vancouver_HQendo_VAFrascal' - Components determined by modelling high-quality p53abn sWGS endometrial samples. (2022)

signatures

String. The signatures themselves. Please name those for which to calculate exposures.
Options:
'signatures/30kb_ovarian/component_by_signature_britroc_aCNs.rds' - The signatures created using the Britroc Ovarian sWGS samples. (Macintyre 2018)
'VAFrascalHQendoVan6sigs' - The signatures created using the high-quality p53abn sWGS endometrial samples.

plot_savepath

(optional) String. The path where to save a sample-by-component heatmap. Please provide a directory.

sigs_savepath

(optional) String. The path where to save a csv of the calculated signature exposures. Please provide a directory.

relative

(optional) Logical. If using relative Copy-Number data as input, set to TRUE. Otherwise, ignore.

refgenome

(optional) String. The reference genome used. (ex. 'hg19', or 'hg38')

data_path

String. The path where to find signatures objects and component models.

Value

A dataframe of the exposures for each sample to each signature.

Details

todo - make examples