Performs some cleaning operations.

These pre-processing steps are:

Replaces chromosome 'X' for '23'.

Removes segments in chromosome 'Y'.

Adds an extra column for the End Position of the segment.

Removes segments whose ratio and ratio_median are -1.

Log2 transforms the ratio and ratio_median columns.

Removes segments with a ratio_median = -Inf.

Usage

CleanBamRatiosFrame(raw_bam_ratios, log_transform)

Arguments

raw_bam_ratios

A data frame: the raw bam ratios file obtained from the bam_ratio.txt file.

log_transform

A boolean: TRUE if you want to log2 transform the ratio_median column. FALSE otherwise. If dealing with Controlfreec data, log_transform should be TRUE. If QDNAseq, log_transform should be FALSE.