Determines if the given segments are LGAs of the given size

Iterates through the segments in segments and firstly checks if the segment meets the definition of an LGA:

Segment and Previous are in the same chromosome arm.

Size of space between Segment and Previous is less than 3Mb. Then, we check that:

Size of Segment >= size_lga

Size of Previous >= size_lga Finally, we check that ratio_median difference between Segment and Previous > THR. If so, then mark segment as an LGA (via 1's in the output frame).

Usage

DetermineNumberOfLGAs(threshold, size_lga, segments)

Arguments

threshold

A float. Threshold in ratio_median difference previously estimated.

size_lga

An integer: size of the LGA to look for.

segments

A data frame: segment data.