Build a QDNAseq object from the DLP output files

In order to do 1-to-1 comparisons between QDNAseq and DLP outputs, it is useful to have outputs in the same format. This function creates a QDNAseq object from the DLP output files and returns it.

Usage

DlpToQDNAseq(
  input_path,
  bin_size = 5e+05,
  genome = "hg19",
  Xincluded = FALSE,
  getfiltered = FALSE
)

Arguments

input_path

(string) The path to the output files generated by DLP. A directory.

bin_size

(integer) The bin size used during copy number calling in base pairs. (default: 500000)

Xincluded

(logical) Whether or not the X chromosome was included in this analysis. (default: FALSE)

genome_used

(string) The reference genome used for alignment. (default: hg19)

getfilter

(logical) Grab the filtered read/segment files from the DLP output directories.

Value

A QDNAseq object.