Remove or set a mask for relative CN bins (QDNAseq object)

Remove or set a mask for relative CN bins

Usage

FilterCNs(
  cnobj,
  filter_by = NULL,
  genome = "hg19",
  minimum_overlap = 5000,
  removebins = FALSE,
  maskgaps = FALSE,
  maskcomCNVs = FALSE,
  maskname = "mask1"
)

Arguments

cnobj

QDNASeq object containing the relative copy number calls as well as the segmented relative copy number calls

filter_by

Optional Dataframe containing genomic regions to filter/mask. One region expected per row.
Expected columns:
chr start end etc. etc.

genome

A character string of the reference genome used. ex "hg38"

minimum_overlap

An integer designating the number minimum number of overlapping bases for a bin to be removed.

removebins

A Logical. If TRUE then remove bins overlapping the indicated regions.

maskgaps

A Logical. If TRUE then remove bins overlapping the centromeric and telomeric regions. Be sure to indicate the correct reference genome.

maskcomCNVs

A Logical. If TRUE then remove bins overlapping commonly copy-number variant regions in this organism. Be sure to indicate the correct reference genome.

maskname

A character string for the name to be used in case of setting a mask. Set this field if providing a DF to the filter_by parameter.

Value

A QDNAseq object filtered or with masks set.

Details

Expects the input of a QDNAseq object containing relative copy-number calls (such as from QDNASeq or WiseCondorX).