Create CN-Signatures

Create copy-number signatures by decomposing a sample-by-component matrix.

Usage

GenerateSignatures(
  sample_by_component,
  nsig,
  seed = 77777,
  nmfalg = "brunet",
  nruns = 1000,
  cores = 4
)

Arguments

sample_by_component

Matrix. Expects a sum-of-posteriors probability matrix.

nsig

Integer. The number of signatures. Likely decided by examining the plot from ChooseNumberSignatures().

seed

Integer. (flexmix param) The random seed to use while modelling.

nmfalg

Character string. (NMF param) The specification of the NMF algorithm.

cores

Integer. (NMF param) The number of cores to use when running NMF.

Value

An NMFfit object. Contains the two matrices post decomposition.

Details

Given a sample-by-component sum-of-posteriors matrix and a matrix rank, decompose into two distinct matrices. One matrix will be a sample-by-signature matrix and the other a signature-by-component matrix. In this package, we refer to the signature-by-component matrix as a set of 'copy-number signatures'.