Plot Mixture Model Components by Signature

This function (MixtureModelPlots), for an indicated CN-Signature (S), returns a list of six ggplots. Each plot visualizes the mixture models used for each feature. All mixture components are plotted, but those elevated for S are shaded.

Usage

MixtureModelPlots(signatures, components, sig_of_interest = 1, threshold = 0.3)

Arguments

signatures

A dataframe with components along the y-axis and signatures along the x.

components

A list of S4 objects belonging to the class 'flexmix'.

sig_of_interest

(optional) A single integer corresponding the signature for which to make plots.
In the absence of any value passed in for this parameter, plots for just the first signature will be returned.

Value

A list of ggplots.