Plot relative CN diversity heatmap

Function to plot the copy number profile from relative copy numbers

Usage

RCNDiversityPlot(
  qdnaseq_obj,
  order_by = NULL,
  cluster = TRUE,
  subset = NULL,
  Xchr = FALSE,
  limits = c(-1.5, 1.5),
  breaks = c(2.5, 1, 0.5, 0, -0.5, -1, -2.5),
  genome = "hg19"
)

Arguments

order_by

optional character vector containing the order the samples in the heatmap should be in

cluster

optional logical Do you want to do a row-wise sort of the heatmap (sort the samples)

subset

optional character vector A subset of samples that you'd like to plot.

limits

vector containing the limits of your colour gradient. Passed to scale_fill_gradientn

breaks

vector containing the limits of your colour gradient. Passed to scale_fill_gradientn

genome

reference genome build

x

dataframe containing the copy number segments in long format: columns should be start, sample id (named sample) and segmented

Value

Heatmap of relative copy number calls

Details

Visualizing relative copy number calls from WiseCondorX and QDNASeq as heatmaps