Make an Alluvial Plot based on maximum signature exposure

SEAlluvialPlot builds a ggplot using the easyalluvial R package. For two sets of exposures on the same set of samples it visualizes how the maximum exposure for one set of signatures is related to the other.

Usage

SEAlluvialPlot(exposA, exposB, orderA = NULL, orderB = NULL)

Arguments

exposA

A matrix of signature exposures. Signatures in rows, samples in columns. Likely the output from CallSignatureExposures().

exposB

A matrix of signature exposures. Signatures in rows, samples in columns.

orderA

(Optional) A character vector of signature names that will enforce an order.
Signature set A is assigned names S.A1 -> S.A#, and B S.B1 -> S.B#
ex. c('S.B2', 'S.B5', 'S.B3', 'S.B4', 'S.B1', 'S.B6')

orderB

(Optional) A character vector of signature names that will enforce an order.
Signature set A is assigned names S.A1 -> S.A#, and B S.B1 -> S.B#
ex. c('S.B2', 'S.B5', 'S.B3', 'S.B4', 'S.B1', 'S.B6') Must be for the same set of same set of samples as provided to exposA. Likely the output from CallSignatureExposures().

Value

A list. A ggplot and its corresponding data matrix.