Create a Stacked Bar-plot of Signature Exposures

Converts signature-per-sample data into a ggplot2 stacked bar plot and returns that object. Samples are sorted by signature exposure (default is the first row). This can be changed to a different row, or a char vector or sample names can be supplied for a custom ordering. The turbo viridis colour-scheme is used.

Usage

StackedExposuresPlot(
  input_matrix,
  user_colours = NULL,
  orderbyrow = 1,
  orderbyname = NULL,
  do_transpose = FALSE
)

Arguments

input_matrix

A dataframe/matrix/datatable with signatures in the rows and samples in the columns..

user_colours

(optional) A vector of colours for the signatures, one per signature.

orderbyrow

(optional) An integer. Single integer corresponding to the row by which to sort.

orderbyname

(optional) NULL or a character vector of sample names.

do_transpose

(optional) Logical. If TRUE then flip bars horizontal.

Value

A ggplot2 object.